+++ src/com/danarossa/database/daointerfaces/IStudentMarkDao.java
@@ -4,10 +4,10 @@ import com.danarossa.entities.StudentMark;
 
 import java.util.List;
 
-public interface IStudentMarkDao extends GenericDao<StudentMark, Long> {
+public interface IStudentMarkDao extends GenericDao<StudentMark, Integer> {
 
-    List<StudentMark> getStudentMarksForRealizedCourse(Long realizedCourseId);
+    List<StudentMark> getStudentMarksForRealizedCourse(Integer realizedCourseId);
 
-    List<StudentMark> getStudentMarksForStudent(Long studentId);
+    List<StudentMark> getStudentMarksForStudent(Integer studentId);
 
 }
