+++ src/com/danarossa/database/daointerfaces/IUserDao.java
@@ -4,6 +4,6 @@ import com.danarossa.entities.User;
 
 import java.util.List;
 
-public interface IUserDao extends GenericDao<User, Long> {
-    List<User> getAllLecturersForStudent(Long studentId);
+public interface IUserDao extends GenericDao<User, Integer> {
+    List<User> getAllLecturersForStudent(Integer studentId);
 }
